+++ dbt/include/fabric/macros/adapters/metadata.sql
@@ -1,12 +1,3 @@
-{% macro use_database_hint() %}
-    {{ return(adapter.dispatch('use_database_hint')()) }}
-{% endmacro %}
-
-{% macro default__use_database_hint() %}{% endmacro %}
-{% macro fabric__use_database_hint() %}
-    {# USE [{{ relation.database }}]; #}
-{% endmacro %}
-
 {% macro information_schema_hints() %}
     {{ return(adapter.dispatch('information_schema_hints')()) }}
 {% endmacro %}
@@ -24,7 +15,7 @@
             name as principal_name,
             principal_id as principal_id
         from
-            sys.database_principals
+            sys.database_principals {{ information_schema_hints() }}
     ),
 
     schemas as (
@@ -33,7 +24,7 @@
             schema_id as schema_id,
             principal_id as principal_id
         from
-            sys.schemas
+            sys.schemas {{ information_schema_hints() }}
     ),
 
     tables as (
@@ -43,7 +34,7 @@
             principal_id as principal_id,
             'BASE TABLE' as table_type
         from
-            sys.tables
+            sys.tables {{ information_schema_hints() }}
     ),
 
     tables_with_metadata as (
@@ -64,7 +55,7 @@
             principal_id as principal_id,
             'VIEW' as table_type
         from
-            sys.views
+            sys.views {{ information_schema_hints() }}
     ),
 
     views_with_metadata as (
@@ -107,7 +98,7 @@
             column_name,
             ordinal_position as column_index,
             data_type as column_type
-        from INFORMATION_SCHEMA.COLUMNS
+        from INFORMATION_SCHEMA.COLUMNS {{ information_schema_hints() }}
 
     )
 
@@ -138,9 +129,9 @@
 
 {% macro fabric__list_schemas(database) %}
   {% call statement('list_schemas', fetch_result=True, auto_begin=False) -%}
-    {{ use_database_hint() }}
+
     select  name as [schema]
-    from sys.schemas
+    from sys.schemas {{ information_schema_hints() }}
   {% endcall %}
   {{ return(load_result('list_schemas').table) }}
 {% endmacro %}
@@ -164,7 +155,7 @@
            else table_type
       end as table_type
 
-    from [{{ schema_relation.database }}].INFORMATION_SCHEMA.TABLES
+    from [{{ schema_relation.database }}].INFORMATION_SCHEMA.TABLES {{ information_schema_hints() }}
     where table_schema like '{{ schema_relation.schema }}'
   {% endcall %}
   {{ return(load_result('list_relations_without_caching').table) }}