+++ dbt/include/fabric/macros/materializations/models/view/create_view_as.sql
@@ -5,8 +5,7 @@
 {% macro fabric__create_view_exec(relation, sql) -%}
 
     {%- set temp_view_sql = sql.replace("'", "''") -%}
-    {{ use_database_hint() }}
-
+    USE [{{ relation.database }}];
     {% set contract_config = config.get('contract') %}
     {% if contract_config.enforced %}
         {{ get_assert_columns_equivalent(sql) }}